Use a PluginHelper to isolate Moka.php
